pyproject.toml
[tool.poetry]
name = "cometa"
version = "0.1.0"
description = ""
authors = ["Christopher Flores"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1.7"
djangorestframework = "3.14.0"
python-decouple = "3.8"
djangorestframework-types = "^0.8.0"
django-cors-headers = "^3.14.0"
turbobus = "==1.0.0a5"
pytz = "^2023.3.post1"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
